requirements.txt
fastapi==0.115.2
uvicorn==0.30.6
pydantic==2.9.2
pydantic-settings==2.5.2
python-dotenv==1.0.1
httpx==0.27.2
pytest==8.3.3
pyyaml==6.0.2
langchain-core==0.3.7
langchain-openai==0.2.1
